Add UpConv2dConnection and image reconstruction
- UpConv2dConnection utilizes upsample in conjunction with convolution
to grow the spatial dimensions by a scale factor. It leverages nearest
neighbor approximation as that should allow for porting to a spiking
network chip, as opposed to bilinear interpolation which is very
unlikely that will easily be able to be transfered (because what is half
a spike?)

- mnist_image_recon - essentially, we try to get the output to fire at
the same frequency as the input spikes into the network. Current version
is 5 layers, have only gotten 3 to run.